requirements.txt
# Core LLM
langchain
langchain-google-genai
langchain-community
langgraph

# PDF Parsing
pymupdf

# Vector Store
chromadb

# Search
google-search-results

# Storage
# sqlite3 is built-in to Python

# UI (later)
streamlit

# Utilities
python-dotenv
pydantic

beautifulsoup4
requests